requirements.txt
flask==3.1.1
tensorflow-cpu==2.19.0
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.6.1
joblib==1.5.1
gunicorn==22.0.0